pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "cryptography>=46.0.5",
    "daphne>=4.2.1",
    "django>=5.2.12",
    "django-celery-beat>=2.9.0",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.15",
    "google-api-python-client>=2.192.0",
    "google-auth-oauthlib>=1.3.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
